new chages in categoryController
code pushed by Sakshi

+++ New_Retail_Management_System_Back/src/main/java/com/example/demo/Model/Product.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class Product {
-	private Long productID;
+	private int  productID;
     private String productName;
     private int productPrice;
     private String pBrand;
